Add abstract trainer
